refactor: Move toolAdapter package under runtime
