+++ internal/runtime/toolAdapter/api/translate.go
