+++ internal/runtime/toolAdapter/script/translator.go
