+++ internal/tools/executor.go
@@ -11,8 +11,8 @@ import (
 	"github.com/pardnchiu/agenvoy/extensions"
 	"github.com/pardnchiu/agenvoy/internal/filesystem"
 	"github.com/pardnchiu/agenvoy/internal/runtime"
-	apiAdapter "github.com/pardnchiu/agenvoy/internal/toolAdapter/api"
-	scriptAdapter "github.com/pardnchiu/agenvoy/internal/toolAdapter/script"
+	apiAdapter "github.com/pardnchiu/agenvoy/internal/runtime/toolAdapter/api"
+	scriptAdapter "github.com/pardnchiu/agenvoy/internal/runtime/toolAdapter/script"
 	toolRegister "github.com/pardnchiu/agenvoy/internal/tools/register"
 	toolTypes "github.com/pardnchiu/agenvoy/internal/tools/types"
 	provider "github.com/pardnchiu/go-llm-router/core"
